requirements.txt
discord.py==2.3.1
Flask==2.3.2
Requests==2.31.0
